+++ application_project/app/src/main/java/com/alyak/detector/navigation/NavGraph.kt
@@ -6,8 +6,8 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.alyak.detector.ui.camera.CameraScreen
-import com.alyak.detector.ui.main.MainScreen
 import com.alyak.detector.ui.map.MapScreen
+import com.alyak.detector.ui.main.MainScreen
 import com.alyak.detector.ui.signIn.SignInScreen
 import com.alyak.detector.ui.signIn.SignInViewModel
 import com.alyak.detector.ui.signUp.SignUpViewModel
