+++ application_project/app/src/main/java/com/alyak/detector/ui/main/DonutChart.kt
@@ -1,4 +1,4 @@
-package com.alyak.detector.ui.family
+package com.alyak.detector.ui.main
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
