+++ application_project/app/src/main/java/com/alyak/detector/ui/main/HistoryBox.kt
@@ -1,4 +1,4 @@
-package com.alyak.detector.ui.family
+package com.alyak.detector.ui.main
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -12,7 +12,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.scale
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.text.font.FontFamily
