+++ application_project/gradle/libs.versions.toml
@@ -1,6 +1,8 @@
 [versions]
 agp = "8.7.2"
 android = "2.12.8"
+appcompatV7 = "28.0.0"
+appcompatVersion = "1.7.1"
 cameraLifecycleVersion = "1.3.0"
 cameraViewVersion = "1.3.0"
 composeBomVersion = "2025.03.01"
@@ -30,6 +32,7 @@ v2Common = "2.20.1"
 
 [libraries]
 android = { module = "com.kakao.maps.open:android", version.ref = "android" }
+androidx-appcompat-v171 = { module = "androidx.appcompat:appcompat", version.ref = "appcompatVersion" }
 androidx-camera-camera2 = { module = "androidx.camera:camera-camera2", version.ref = "cameraLifecycleVersion" }
 androidx-camera-core = { module = "androidx.camera:camera-core", version.ref = "cameraLifecycleVersion" }
 androidx-camera-lifecycle-v130 = { module = "androidx.camera:camera-lifecycle", version.ref = "cameraLifecycleVersion" }
@@ -41,6 +44,7 @@ androidx-hilt-navigation-compose = { module = "androidx.hilt:hilt-navigation-com
 androidx-mateiral3 = { module = "androidx.compose.material3:mateiral3" }
 androidx-material-icons-extended = { module = "androidx.compose.material:material-icons-extended", version.ref = "materialIconsExtended" }
 androidx-navigation-compose = { module = "androidx.navigation:navigation-compose", version.ref = "navigationCompose" }
+appcompat-v7 = { module = "com.android.support:appcompat-v7", version.ref = "appcompatV7" }
 converter-gson = { module = "com.squareup.retrofit2:converter-gson", version.ref = "converterGson" }
 junit = { group = "junit", name = "junit", version.ref = "junit" }
 androidx-junit = { group = "androidx.test.ext", name = "junit", version.ref = "junitVersion" }
